Allow greater variability for handle movement
Add `travel` setting to view which controls how far the handle can move from the center of the base. The
default value of 1.0 gives previous behavior with max travel resulting in the handle center lying on the
circumference of the base. Values greater than 1.0 let the handle travel beyond the base circumference
and values less than 1.0 restrict travel to within the base area.